Simplex-DW/requirements/requirements.txt
apache-airflow-providers-snowflake==3.3.0
apache-airflow-providers-docker==3.7.5
snowflake-connector-python==2.8.1
snowflake-sqlalchemy==1.4.3
pymysql
mysql
oscrypto==1.3.0
pyOpenSSL==22.0.0
pycryptodomex==3.15.0
apache-airflow == 2.4.3
apache-airflow-providers-postgres==5.2.2
apache-airflow-providers-slack==6.0.0
apache-airflow-providers-sftp==4.1.0
fsspec==2022.11.0
httpx==0.23.0
psycopg2==2.9.5
python-dateutil==2.8.2
SQLAlchemy==1.4.27
pandas==1.5.1
simple_salesforce==1.12.3
aiohttp==3.8.3
pyarrow==8.0.0
numpy==1.26.2
matplotlib==3.8.2